requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
psycopg
alembic
pydantic-settings
